logging v2

+++ src/common/types.rs
@@ -1,2 +1,2 @@
-pub type CalcErr = (String, Option<u16>);
+pub type CalcErr = (String, Option<usize>);
 pub type CalcResult<T> = Result<T, CalcErr>;
